requirements.txt
# Core RAG components
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# LLM provider
groq>=0.4.0  # Groq API for fast inference

# Web API
fastapi>=0.111.0
uvicorn>=0.30.0
python-dotenv>=1.0.0

# Optional: For Jupyter notebooks
# jupyter>=1.0.0
# matplotlib>=3.7.0
# seaborn>=0.12.0
